Removed tyhe typecasts a 'long long' in unittests. Fixed a bug about expirationTime == -1

+++ test/unittests/mongoBackend/mongoGetSubscriptions_test.cpp
@@ -75,7 +75,7 @@ static void prepareDatabaseV1Subs(void) {
     DBClientBase* connection = getMongoConnection();
 
     BSONObj sub1 = BSON("_id" << OID(SUB_OID1) <<
-                        "expiration" << (long long) 10000000 <<                        
+                        "expiration" << 10000000 <<                        
                         "reference" << "http://notify1.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "type" << "T1" << "isPattern" << "false")) <<
                         "attrs" << BSONArray() <<
@@ -86,9 +86,9 @@ static void prepareDatabaseV1Subs(void) {
                         );
 
     BSONObj sub2 = BSON("_id" << OID(SUB_OID2) <<
-                        "expiration" << (long long) 25000000 <<
+                        "expiration" << 25000000 <<
                         "lastNotification" << 20000000 <<
-                        "count" << (long long) 24 <<
+                        "count" << 24 <<
                         "reference" << "http://notify2.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E.*" << "type" << "T2" << "isPattern" << "true")) <<
                         "attrs" << BSON_ARRAY("A1" << "A2") <<
@@ -96,11 +96,11 @@ static void prepareDatabaseV1Subs(void) {
                                                        "type" << "ONCHANGE" <<
                                                        "value" << BSON_ARRAY("AX2" << "AY2")
                                                        )) <<
-                        "throttling" << (long long) 5
+                        "throttling" << 5.0
                         );
 
     BSONObj sub3 = BSON("_id" << OID(SUB_OID3) <<
-                        "expiration" << (long long) 20000000 <<
+                        "expiration" << 20000000 <<
                         "lastNotification" << 25000000 <<
                         "reference" << "http://notify2.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E.*" << "type" << "T2" << "isPattern" << "true")) <<
