+++ test/unittests/mongoBackend/mongoUpdateContext_withOnchangeSubscriptions_test.cpp
@@ -233,7 +233,7 @@ static void prepareDatabase(bool useSubCache = true)
                     );
 
   BSONObj sub1 = BSON("_id" << OID("51307b66f481db11bf860001") <<
-                      "expiration" << (long long) 1500000000 <<
+                      "expiration" << 1500000000 <<
                       "lastNotification" << 20000000 <<
                       "reference" << "http://notify1.me" <<
                       "entities" << BSON_ARRAY(BSON("id" << "E1" << "type" << "T1" << "isPattern" << "false")) <<
@@ -245,7 +245,7 @@ static void prepareDatabase(bool useSubCache = true)
                       );
 
   BSONObj sub2 = BSON("_id" << OID("51307b66f481db11bf860002") <<
-                      "expiration" << (long long) 2000000000 <<
+                      "expiration" << 2000000000 <<
                       "lastNotification" << 30000000 <<
                       "reference" << "http://notify2.me" <<
                       "entities" << BSON_ARRAY(BSON("id" << "E2" << "type" << "T2" << "isPattern" << "false")) <<
@@ -269,7 +269,7 @@ static void prepareDatabase(bool useSubCache = true)
                       );
 
   BSONObj sub3 = BSON("_id" << OID("51307b66f481db11bf860003") <<
-                      "expiration" << (long long) 1500000000 <<
+                      "expiration" << 1500000000 <<
                       "lastNotification" << 20000000 <<
                       "reference" << "http://notify3.me" <<
                       "entities" << BSON_ARRAY(BSON("id" << "E[1-2]" << "type" << "T" << "isPattern" << "true")) <<
@@ -349,7 +349,7 @@ static void prepareDatabaseWithNoTypeSubscriptions(void)
                       );
 
     BSONObj sub4 = BSON("_id" << OID("51307b66f481db11bf860004") <<
-                        "expiration" << (long long) 1500000000 <<
+                        "expiration" << 1500000000 <<
                         "lastNotification" << 20000000 <<
                         "reference" << "http://notify4.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "isPattern" << "false")) <<
@@ -361,7 +361,7 @@ static void prepareDatabaseWithNoTypeSubscriptions(void)
                         );
 
     BSONObj sub5 = BSON("_id" << OID("51307b66f481db11bf860005") <<
-                        "expiration" << (long long) 1500000000 <<
+                        "expiration" << 1500000000 <<
                         "lastNotification" << 20000000 <<
                         "reference" << "http://notify5.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E[2-3]" << "isPattern" << "true")) <<
